+++ src/spec/AnimationSpec.ts
@@ -547,4 +547,124 @@ describe('A Graphics Animation', () => {
     anim.tick(expectedFrameDuration, 2);
     expect(anim.currentFrame).toBe(anim.frames[0]);
   });
+
+  it('has a current time left in a frame', () => {
+
+    const rect = new ex.Rectangle({
+      width: 100,
+      height: 100,
+      color: ex.Color.Blue
+    });
+    const frames: ex.Frame[] = [
+      {
+        graphic: rect,
+        duration: 100
+      },
+      {
+        graphic: rect,
+        duration: 100
+      }
+    ];
+    const anim = new ex.Animation({
+      frames: frames
+    });
+    anim.play();
+    expect(anim.currentFrameIndex).toBe(0);
+    expect(anim.currentFrameTimeLeft).toBe(100);
+    anim.tick(10, 1);
+    expect(anim.currentFrameIndex).toBe(0);
+    expect(anim.currentFrameTimeLeft).toBe(90);
+    anim.tick(10, 2);
+    expect(anim.currentFrameIndex).toBe(0);
+    expect(anim.currentFrameTimeLeft).toBe(80);
+    anim.tick(80, 3);
+    expect(anim.currentFrameIndex).toBe(1);
+    expect(anim.currentFrameTimeLeft).toBe(100);
+  });
+
+  it('can go to a frame with an overridden duration', () => {
+    const rect = new ex.Rectangle({
+      width: 100,
+      height: 100,
+      color: ex.Color.Blue
+    });
+    const frames: ex.Frame[] = [
+      {
+        graphic: rect,
+        duration: 100
+      },
+      {
+        graphic: rect,
+        duration: 100
+      }
+    ];
+    const anim = new ex.Animation({
+      frames: frames
+    });
+    anim.play();
+    expect(anim.currentFrameIndex).toBe(0);
+    expect(anim.currentFrameTimeLeft).toBe(100);
+
+    anim.goToFrame(1, 50);
+    expect(anim.currentFrameIndex).toBe(1);
+    expect(anim.currentFrameTimeLeft).toBe(50);
+  });
+
+  it('can adjust playback speed', () => {
+    const rect = new ex.Rectangle({
+      width: 100,
+      height: 100,
+      color: ex.Color.Blue
+    });
+    const frames: ex.Frame[] = [
+      {
+        graphic: rect,
+        duration: 100
+      },
+      {
+        graphic: rect,
+        duration: 100
+      }
+    ];
+    const anim = new ex.Animation({
+      frames: frames,
+      speed: 2
+    });
+    anim.play();
+    expect(anim.currentFrameIndex).toBe(0);
+    expect(anim.currentFrameTimeLeft).toBe(100);
+
+    anim.tick(20, 1);
+    expect(anim.currentFrameIndex).toBe(0);
+    expect(anim.currentFrameTimeLeft).toBe(60);
+  });
+
+  it('can adjust playback speed (only positive', () => {
+    const rect = new ex.Rectangle({
+      width: 100,
+      height: 100,
+      color: ex.Color.Blue
+    });
+    const frames: ex.Frame[] = [
+      {
+        graphic: rect,
+        duration: 100
+      },
+      {
+        graphic: rect,
+        duration: 100
+      }
+    ];
+    const anim = new ex.Animation({
+      frames: frames
+    });
+    anim.speed = -100;
+    expect(anim.speed).toBe(100);
+
+    anim.speed = 0;
+    expect(anim.speed).toBe(0);
+
+    anim.speed = 100;
+    expect(anim.speed).toBe(100);
+  });
 });